Fixed problems recording and playing back audio
This example file will record 5 seconds of audio.
Unfortunately, there is a segfault at the very end
of the program while I am attempting to clean up.
It is worth noting that in order to get this program
to work at all. I had to stop the pulseaudio daemon
that was running on my machine.